ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
prophet==1.1.5
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
kaggle==1.5.16
